test: Migrate geo-utils test
- remove old test & mockup files
