--- tests/unit/mockups/geo-utils.js
@@ -1,19 +1,0 @@
-
-const longLatCoordinates = [
-  [ 8.609088, 49.469343, 106 ],
-  [ 8.609031, 49.469324, 105 ],
-  [ 8.608745, 49.469206, 104.5 ],
-  [ 8.608382, 49.469103, 104.6 ],
-  [ 8.607994, 49.469053, 104.7 ]
-]
-const latLongCoordinates = [
-  [ 49.469343, 8.609088, 106 ],
-  [ 49.469324, 8.609031, 105 ],
-  [ 49.469206, 8.608745, 104.5 ],
-  [ 49.469103, 8.608382, 104.6 ],
-  [ 49.469053, 8.607994, 104.7 ]
-]
-export default {
-  longLatCoordinates,
-  latLongCoordinates
-}
